+++ Invalidator.d.ts
@@ -1,0 +1,1 @@
+export declare type Invalidator<I extends unknown> = (value?: I) => void;