+++ Invalidator.ts
@@ -1,0 +1,1 @@
+export type Invalidator<I extends unknown> = (value?: I) => void