+++ StartStopNotifier.ts
@@ -1,0 +1,3 @@
+import type { Subscriber } from './Subscriber'
+import type { Unsubscriber } from './Unsubscriber'
+export type StartStopNotifier<I extends unknown> = (set:Subscriber<I>)=>Unsubscriber|void