+++ Subscriber.ts
@@ -1,0 +1,1 @@
+export type Subscriber<I extends unknown> = (value: I) => void