+++ Updater.d.ts
@@ -1,0 +1,1 @@
+export declare type Updater<I extends unknown> = (value: I) => I;