+++ Updater.ts
@@ -1,0 +1,1 @@
+export type Updater<I extends unknown> = (value: I) => I