+++ _readable_set_a1.d.ts
@@ -1,2 +1,2 @@
-export declare function _readable_set_a1<I>(initial: any): (_readable_set_a1_set_type<I> | import("svelte/types/runtime/store").Readable<I>)[];
-export declare type _readable_set_a1_set_type<I> = (val: I) => void;
+export declare function _readable_set_a1<I extends unknown>(initial: any): (_readable_set_a1_set_type<I> | import("svelte/types/runtime/store").Readable<I>)[];
+export declare type _readable_set_a1_set_type<I extends unknown> = (val: I) => void;