+++ _set_store.ts
@@ -3,7 +3,7 @@ import { set } from './set'
  * Returns a function to set the given store using the value returned by `setter`.
  * This is useful in conjunction with [subscribe](#subscribe).
  */
-export function _set_store<I>(store, setter = (v:I)=>v) {
+export function _set_store<I extends unknown>(store, setter = (v:I)=>v) {
 	return (val: I)=>
 		set<I>(store,
 			typeof setter === 'function'