+++ derived__async.ts
@@ -4,20 +4,24 @@ import { subscribe } from './subscribe'
 import type { Unsubscriber } from './Unsubscriber'
 import type { Stores } from './Stores'
 import type { Readable } from './lib'
+import type { Subscriber } from './Subscriber'
 /**
  * Creates a Readable store that derives it's value from a async function.
  * @see derived__store
  */
-export function derived__async<I>(in_stores:Stores<I>, fn, initial_value = null) {
+export function derived__async<I extends unknown>(
+	in_stores:Stores<I>,
+	fn: (values: I | I[], set: Subscriber<I>) => Promise<I>,
+	initial_value = null as I
+) {
 	const single = !Array.isArray(in_stores)
 	const stores = (single ? [in_stores] : in_stores) as Readable<I>[]
 	const auto = fn.length < 2
-	let value = {}
+	let value: I
 	return readable(initial_value, set=>{
 		let inited = false
 		const values = [] as I[]
 		let pending = 0
-		// @ts-ignore
 		async function sync():Promise<void> {
 			if (pending) return
 			const result = await fn(single ? values[0] : values, set)