+++ set.d.ts
@@ -1,8 +1,5 @@
 /**
  * Calls set on the given store with the given val
- * @param {Writable} store
- * @param val
- * @returns {void}
  */
 import type { Writable } from 'svelte/store';
-export declare function set<I>(store: Writable<I>, val: any): void;
+export declare function set<I extends unknown>(store: Writable<I>, val: I): void;