+++ set.ts
@@ -1,10 +1,7 @@
 /**
  * Calls set on the given store with the given val
- * @param {Writable} store
- * @param val
- * @returns {void}
  */
 import type { Writable } from 'svelte/store'
-export function set<I>(store: Writable<I>, val) {
+export function set<I extends unknown>(store: Writable<I>, val: I) {
 	store.set(val)
 }