+++ subscribe.d.ts
@@ -1,9 +1,8 @@
-import type { Readable } from 'svelte/store';
+import type { Readable } from './lib';
+import type { Subscriber } from './Subscriber';
+import type { Invalidator } from './Invalidator';
+import type { Unsubscriber } from './Unsubscriber';
 /**
  * Delegates to store.subscribe
- * @param {Readable} store
- * @param {function} run
- * @param {function} [invalidate]
- * @returns {Unsubscriber}
  */
-export declare function subscribe<I>(store: Readable<I>, run: (I: any) => void, invalidate?: (I: any) => void): () => void;
+export declare function subscribe<I extends unknown>(store: Readable<I>, run: Subscriber<I>, invalidate?: Invalidator<I>): Unsubscriber;