+++ subscribe.ts
@@ -1,15 +1,14 @@
-import type { Readable } from 'svelte/store'
+import type { Readable } from './lib'
+import type { Subscriber } from './Subscriber'
+import type { Invalidator } from './Invalidator'
+import type { Unsubscriber } from './Unsubscriber'
 /**
  * Delegates to store.subscribe
- * @param {Readable} store
- * @param {function} run
- * @param {function} [invalidate]
- * @returns {Unsubscriber}
  */
-export function subscribe<I>(
+export function subscribe<I extends unknown>(
 	store:Readable<I>,
-	run:(I)=>void,
-	invalidate?:(I)=>void
+	run:Subscriber<I>,
+	invalidate?:Invalidator<I>
 ) {
-	return store.subscribe(run, invalidate)
+	return store.subscribe(run, invalidate) as Unsubscriber
 }