+++ subscribe__change__once.d.ts
@@ -1,7 +1,6 @@
+import type { Readable } from './lib';
+import type { Subscriber } from './Subscriber';
 /**
  * Calls the given fn the next time the value of the store changes, then unsubscribes.
- * @param {Readable} store
- * @param {function} fn
- * @returns {Unsubscriber}
  */
-export declare function subscribe__change__once<T>(store: any, fn: any): () => void;
+export declare function subscribe__change__once<I extends unknown>(store: Readable<I>, fn: Subscriber<I>): import("./Unsubscriber").Unsubscriber;