+++ subscribe__change__once.ts
@@ -1,16 +1,16 @@
 import { subscribe__noinit } from './subscribe__noinit'
+import type { Readable } from './lib'
+import type { Subscriber } from './Subscriber'
 /**
  * Calls the given fn the next time the value of the store changes, then unsubscribes.
- * @param {Readable} store
- * @param {function} fn
- * @returns {Unsubscriber}
  */
-export function subscribe__change__once<T>(store, fn) {
-	const unsubscribe = subscribe__noinit<T>(store, (...arg_a1)=>{
-		// @ts-ignore
-		const __ = fn(...arg_a1)
+export function subscribe__change__once<I extends unknown>(
+	store: Readable<I>, fn: Subscriber<I>
+) {
+	const unsubscribe = subscribe__noinit<I>(store, (value=>{
+		const out = fn(value)
 		unsubscribe()
-		return __
-	})
+		return out
+	}) as Subscriber<I>)
 	return unsubscribe
 }