+++ subscribe__noinit.d.ts
@@ -1,7 +1,5 @@
+import type { Readable } from './lib';
 /**
  * Subscribes the fn to store but does not have the initial call.
- * @param {Readable} store
- * @param {function} fn
- * @returns {function: void}
  */
-export declare function subscribe__noinit<T>(store: any, fn: any): () => void;
+export declare function subscribe__noinit<I extends unknown>(store: Readable<I>, fn: any): import("./Unsubscriber").Unsubscriber;